move html stuff out of ctrl.ts
