+++ ui/ceval/src/ctrl.ts
@@ -1,7 +1,7 @@
 import throttle from 'common/throttle';
 import { Engines } from './engines/engines';
 import { CevalOpts, CevalState, CevalEngine, Work, Step, Hovering, PvBoard, Started } from './types';
-import { sanIrreversible } from './util';
+import { sanIrreversible, showEngineError } from './util';
 import { defaultPosition, setupPosition } from 'chessops/variant';
 import { parseFen } from 'chessops/fen';
 import { lichessRules } from 'chessops/compat';
@@ -11,7 +11,6 @@ import { hasFeature } from 'common/device';
 import { Result } from '@badrap/result';
 import { storedIntProp } from 'common/storage';
 import { Rules } from 'chessops';
-import { type Dialog, domDialog } from 'common/dialog';
 
 const cevalDisabledSentinel = '1';
 
@@ -238,24 +237,7 @@ export default class CevalCtrl {
   };
 
   engineFailed(msg: string) {
-    domDialog({
-      class: 'engine-error',
-      htmlText:
-        `<h2>${lichess.escapeHtml(this.engines.active?.name ?? 'Engine')} <bad>error</bad></h2>` +
-        `<pre tabindex="0" class="err">${lichess.escapeHtml(msg)}</pre><h2>Things to try</h2><ul>` +
-        '<li>Decrease memory slider in engine settings</li><li>Clear site settings for lichess.org</li>' +
-        '<li>Select another engine</li><li>Update your browser</li></ul>',
-    }).then((dlg: Dialog) => {
-      const select = () =>
-        setTimeout(() => {
-          const range = document.createRange();
-          range.selectNodeContents(dlg.view.querySelector('.err')!);
-          window.getSelection()?.removeAllRanges();
-          window.getSelection()?.addRange(range);
-        }, 0);
-      dlg.view.querySelector('.err')?.addEventListener('focus', select);
-      dlg.showModal();
-    });
+    showEngineError(this.engines.active?.name ?? 'Engine', msg);
   }
 
   cacheable() {
